+++ src/helpers/get-domain.helper.ts
@@ -1,0 +1,6 @@
+export function getDomain(url: string): string {
+  return url
+    .replace(/https?:\/\//, '')
+    .split('/')[0]
+    .split(':')[0];
+}